src/requirements.txt
kafka-python
sqlalchemy
nltk
newspaper3k